+++ packages/cli/src/types.d.ts
@@ -26,10 +26,6 @@ declare module 'rollup-plugin-image-files' {
   export default function image(options?: any): any;
 }
 
-declare module '@svgr/rollup' {
-  export default function svgr(options?: any): any;
-}
-
 declare module 'rollup/parseAst' {
   export function parseAst(code: string): any;
   export function parseAstAsync(code: string): any;