+++ packages/frontend-app-api/src/wiring/createSpecializedApp.test.tsx
@@ -144,8 +144,8 @@ describe('createSpecializedApp', () => {
               ],
             }),
             ApiBlueprint.make({
-              params: define =>
-                define({
+              params: defineParams =>
+                defineParams({
                   api: featureFlagsApiRef,
                   deps: {},
                   factory: () =>
@@ -253,8 +253,8 @@ describe('createSpecializedApp', () => {
           pluginId: 'first',
           extensions: [
             ApiBlueprint.make({
-              params: define =>
-                define({
+              params: defineParams =>
+                defineParams({
                   api: analyticsApiRef,
                   deps: {},
                   factory: () => {
@@ -294,8 +294,8 @@ describe('createSpecializedApp', () => {
               },
             }),
             ApiBlueprint.make({
-              params: define =>
-                define({
+              params: defineParams =>
+                defineParams({
                   api: analyticsApiRef,
                   deps: {},
                   factory: mockAnalyticsApi,
@@ -313,10 +313,12 @@ describe('createSpecializedApp', () => {
 
   it('should use provided apis', async () => {
     const app = createSpecializedApp({
-      apis: TestApiRegistry.from([
-        configApiRef,
-        new ConfigReader({ anything: 'config' }),
-      ]),
+      advanced: {
+        apis: TestApiRegistry.from([
+          configApiRef,
+          new ConfigReader({ anything: 'config' }),
+        ]),
+      },
       features: [
         createFrontendPlugin({
           pluginId: 'test',
@@ -639,28 +641,30 @@ describe('createSpecializedApp', () => {
           ],
         }),
       ],
-      extensionFactoryMiddleware: [
-        function* middleware(originalFactory, { config }) {
-          const result = originalFactory({
-            config: config && { text: `1-${config.text}` },
-          });
-          yield* result;
-          const el = result.get(textDataRef);
-          if (el) {
-            yield textDataRef(`${el}-1`);
-          }
-        },
-        function* middleware(originalFactory, { config }) {
-          const result = originalFactory({
-            config: config && { text: `2-${config.text}` },
-          });
-          yield* result;
-          const el = result.get(textDataRef);
-          if (el) {
-            yield textDataRef(`${el}-2`);
-          }
-        },
-      ],
+      advanced: {
+        extensionFactoryMiddleware: [
+          function* middleware(originalFactory, { config }) {
+            const result = originalFactory({
+              config: config && { text: `1-${config.text}` },
+            });
+            yield* result;
+            const el = result.get(textDataRef);
+            if (el) {
+              yield textDataRef(`${el}-1`);
+            }
+          },
+          function* middleware(originalFactory, { config }) {
+            const result = originalFactory({
+              config: config && { text: `2-${config.text}` },
+            });
+            yield* result;
+            const el = result.get(textDataRef);
+            if (el) {
+              yield textDataRef(`${el}-2`);
+            }
+          },
+        ],
+      },
     });
 
     const root = app.tree.root.instance!.getData(
@@ -691,7 +695,7 @@ describe('createSpecializedApp', () => {
 
       await expect(plugin.info()).rejects.toThrow(errorMsg);
 
-      const installedPlugin = app.tree.nodes.get('test')?.spec.source;
+      const installedPlugin = app.tree.nodes.get('test')?.spec.plugin;
       expect(installedPlugin).toBeDefined();
       const info = await installedPlugin?.info();
       expect(info).toEqual({});
@@ -707,7 +711,7 @@ describe('createSpecializedApp', () => {
       });
 
       const app = createSpecializedApp({ features: [plugin] });
-      const info = await app.tree.nodes.get('test')?.spec.source?.info();
+      const info = await app.tree.nodes.get('test')?.spec.plugin?.info();
       expect(info).toMatchObject({
         packageName: '@backstage/frontend-app-api',
       });
@@ -730,7 +734,7 @@ describe('createSpecializedApp', () => {
       });
 
       const app = createSpecializedApp({ features: [overriddenPlugin] });
-      const info = await app.tree.nodes.get('test')?.spec.source?.info();
+      const info = await app.tree.nodes.get('test')?.spec.plugin?.info();
       expect(info).toMatchObject({
         packageName: 'test-override',
       });
@@ -754,7 +758,7 @@ describe('createSpecializedApp', () => {
       });
 
       const app = createSpecializedApp({ features: [plugin] });
-      const info = await app.tree.nodes.get('test')?.spec.source?.info();
+      const info = await app.tree.nodes.get('test')?.spec.plugin?.info();
       expect(info).toEqual({
         packageName: '@backstage/frontend-app-api',
         version: expect.any(String),
@@ -774,15 +778,17 @@ describe('createSpecializedApp', () => {
 
       const app = createSpecializedApp({
         features: [plugin],
-        async pluginInfoResolver(ctx) {
-          const { info } = await ctx.defaultResolver({
-            packageJson: await ctx.packageJson(),
-            manifest: await ctx.manifest(),
-          });
-          return { info: { packageName: `decorated:${info.packageName}` } };
+        advanced: {
+          pluginInfoResolver: async ctx => {
+            const { info } = await ctx.defaultResolver({
+              packageJson: await ctx.packageJson(),
+              manifest: await ctx.manifest(),
+            });
+            return { info: { packageName: `decorated:${info.packageName}` } };
+          },
         },
       });
-      const info = await app.tree.nodes.get('test')?.spec.source?.info();
+      const info = await app.tree.nodes.get('test')?.spec.plugin?.info();
       expect(info).toEqual({
         packageName: 'decorated:@backstage/frontend-app-api',
       });