+++ packages/frontend-plugin-api/src/analytics/index.ts
@@ -15,5 +15,5 @@
  */
 
 export { AnalyticsContext } from './AnalyticsContext';
-export type { AnalyticsContextValue, CommonAnalyticsContext } from './types';
+export type { AnalyticsContextValue } from './types';
 export { useAnalytics } from './useAnalytics';