+++ plugins/org/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/plugin-org
 
+## 0.6.42-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/frontend-plugin-api@0.11.0-next.1
+  - @backstage/core-compat-api@0.5.0-next.2
+  - @backstage/plugin-catalog-react@1.20.0-next.2
+  - @backstage/core-components@0.17.5-next.1
+  - @backstage/catalog-model@1.7.5
+  - @backstage/core-plugin-api@1.10.9
+  - @backstage/plugin-catalog-common@1.1.5
+
 ## 0.6.42-next.1
 
 ### Patch Changes